component 分けた
